requirements.txt
fastapi
uvicorn
elasticsearch==7.13.2
langchain-openai
python-dotenv
asyncpg
pydantic
numpy<2.0